requirements.txt
fastapi==0.65.1
uvicorn==0.13.4
pygithub==1.55

